refactor: move editing event into hook
