+++ apps/dashboard/src/app/(dashboard)/event/[id]/edit-card.tsx
@@ -4,33 +4,14 @@ import { useEventWriteForm } from "../write-form"
 import { EventWriteSchema } from "@dotkomonline/types"
 import { useEventDetailsContext } from "./provider"
 import { useQueryNotification } from "../../../notifications"
+import { useEditEvent } from "../../../../modules/event/use-edit-event"
 
 export const EventEditCard: FC = () => {
   const { event } = useEventDetailsContext()
-  const notification = useQueryNotification()
-  const utils = trpc.useContext()
-  const edit = trpc.event.edit.useMutation({
-    onSuccess: (data) => {
-      notification.complete({
-        title: "Arrangement oppdatert",
-        message: `Arrangementet "${data.title}" har blitt oppdatert.`,
-      })
-      utils.event.all.invalidate()
-    },
-    onError: (err) => {
-      notification.fail({
-        title: "Feil oppsto",
-        message: `En feil oppsto under oppdatering av arrangementet: ${err.toString()}.`,
-      })
-    },
-  })
+  const edit = useEditEvent()
   const FormComponent = useEventWriteForm({
     label: "Oppdater arrangement",
     onSubmit: (data) => {
-      notification.loading({
-        title: "Oppdaterer arrangement...",
-        message: "Arrangementet blir oppdatert.",
-      })
       const result = EventWriteSchema.required({ id: true }).parse(data)
       edit.mutate(result)
     },
